Refactor EmailService to interface
